Updating routes to include new pages.
