+++ pages/library-card/account/index.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function AccountPage() {
+  return (
+    <div>
+      <h2>Step 5 of 6: Create Your Account</h2>
+    </div>
+  );
+}
+
+export default AccountPage;
