+++ pages/library-card/congrats/index.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function CongratsPage() {
+  return (
+    <div>
+      <h2>Congratulations! You now have an NPL digital library card.</h2>
+    </div>
+  );
+}
+
+export default CongratsPage;
