+++ pages/library-card/location/index.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import LibraryCardForm from "../../../src/components/LibraryCardForm";
+
+function LocationPage() {
+  return (
+    <div>
+      <h2>Step 1 of 6: Location</h2>
+      <LibraryCardForm />
+    </div>
+  );
+}
+
+export default LocationPage;
