+++ pages/library-card/personal/index.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function PersonalInformationPage() {
+  return (
+    <div>
+      <h2>Step 2 of 6: Personal Information</h2>
+    </div>
+  );
+}
+
+export default PersonalInformationPage;
